Refactoring/consolidation of source code
I updated the code so that unnecessary parts that didn't function were removed. I also improved the filename collision renaming section by using an incremental suffix instead of adding a plain prefix that kept getting added every time the condition ran, which could result in long filenames. However, one more bug was found by me. The bug occurs when trying to rename files to a filename that already existed in the directory before the script ran. This causes filename collision that is currently not handled. A fix will be coming shortly!