Large overhaul for source code
I fixed the bug mentioned in my prior commit for this file. In doing so, I found an elegant way to handle duplicate filenames without using dictionaries or sets. I simply let the exception at hand occur and handle it properly in a loop. In addition, I consolidated more code and removed the last for loop as it was not necessary. Lastly, I renamed the temp_string variables in the my_rename() function to something more suitable and did it so that no unnecessary variables were declared, which is more efficient than my previous implementation.