Code modularization
I modularized the source code by putting the last section of the code (which renamed the files and printed out messages) into its own function. I also made a main() function and updated the code so that it could be imported as a module in addition to being used as a standalone script. Lastly, I specified python3 on line 1 and the utf-8 character encoding set on line 2 for this program.